Refactored the store page, placed files in a new folder for organization, Store page has new look
