+++ app/navigation/StoreNavigator.js
@@ -1,16 +1,23 @@
 import React from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
-import AccessoriesStore from '../screens/AccessoriesStoreScreen';
-import ToyStore from '../screens/ToyStoreScreen';
-import TreatStore from '../screens/TreatStoreScreen';
-import StoreScreen from '../screens/StoreScreen';
-import GroomingStore from '../screens/GroomingStore';
-import { NavigationContainer } from '@react-navigation/native';
+
+import AccessoriesStore from '../screens/storeScreens/AccessoriesStoreScreen';
+import colors from '../config/colors';
+import GroomingStore from '../screens/storeScreens/GroomingStore';
+import StoreScreen from '../screens/storeScreens/StoreScreen';
+import ToyStore from '../screens/storeScreens/ToyStoreScreen';
+import TreatStore from '../screens/storeScreens/TreatStoreScreen';
+// import { NavigationContainer } from '@react-navigation/native';
 
 const Stack = createStackNavigator();
 
 const StoreNavigator = () => (
-  <Stack.Navigator>
+  <Stack.Navigator
+    screenOptions={{
+      headerStyle: { backgroundColor: colors.purple },
+      headerTintColor: 'white',
+    }}
+  >
     <Stack.Screen name="Store" component={StoreScreen} />
     <Stack.Screen name="AccessoriesStore" component={AccessoriesStore} />
     <Stack.Screen name="ToyStore" component={ToyStore} />
