+++ app/screens/storeScreens/TreatStoreScreen.js
@@ -8,12 +8,13 @@ import {
   ScrollView,
 } from 'react-native';
 import { collection, doc, getDocs } from 'firebase/firestore';
-import { db } from '../../firebase/firebase-config';
+import { db } from '../../../firebase/firebase-config';
 import * as Linking from 'expo-linking';
-import colors from '../config/colors';
-import Screen from '../components/Screen';
+import colors from '../../config/colors';
+import Screen from '../../components/Screen';
+import { ListItem } from '../../components/lists';
 
-function ToyStore() {
+function TreatStore() {
   const [products, setProducts] = useState([]);
 
   const storeCollectionRef = collection(db, 'store');
@@ -24,7 +25,9 @@ function ToyStore() {
         ...document.data(),
         id: document.id,
       }));
-      const filteredData = mappedData.filter((post) => post.category === 'toy');
+      const filteredData = mappedData.filter(
+        (post) => post.category === 'treat'
+      );
       setProducts(filteredData);
     };
     getItems();
@@ -38,7 +41,7 @@ function ToyStore() {
     <Screen style={styles.screen}>
       <ScrollView>
         <View>
-          <Text style={styles.label}>Toy Recommendations:</Text>
+          <Text style={styles.label}>Treat Recommendations:</Text>
           {products ? (
             products.map((item, i) => (
               <View key={i}>
@@ -89,4 +92,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default ToyStore;
+export default TreatStore;
